Extract the RN-free half of ScrollView interop from ScrollViewResponderInterceptor
